pyproject.toml
[tool.poetry]
package-mode = false
name = "dt-fileviewer-web"
version = "0.1.0"
description = "Text File viewer web application"
authors = ["Al DAmico <JavaWiz1@hotmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
aiofiles = "^24.1.0"
dt-foundation = "^0.1.3"
websockets = "^13.1"
jinja2 = "^3.1.4"
requests = "^2.32.3"
python-multipart = "^0.0.16"
python-dateutil = "^2.9.0.post0"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "piwheels"
url = "https://www.piwheels.org/simple"
priority = "supplemental"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
